+++ src/screens/App/EncerramentoConta/index.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { View, Text, StyleSheet, Image, Dimensions } from 'react-native';
-import CustomButton from '../components/CustomButton'; 
+import CustomButton from '@components/CustomButton'; 
 import { useNavigation } from '@react-navigation/native';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../types/types';
+import { RootStackParamList } from '@typings/types';
+import styles from './styles';
 
 const { width, height } = Dimensions.get('window');
 
@@ -35,29 +36,3 @@ export default function EncerramentoConta({ navigation }: Props) {
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingHorizontal: 30,
-    paddingTop: 40,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  successMessage: {
-    fontSize: width * 0.06,
-    fontWeight: 'bold',
-    marginBottom: height * 0.02,
-  },
-  image: {
-    width: 305, 
-    height: 298, 
-    marginBottom: height * 0.02,
-    marginTop: 20,
-  },
-  customButton: {
-    marginTop: height * 0.04,
-    marginBottom: height * 0.1,
-    alignSelf: 'center',
-  },
-});
