+++ src/screens/App/Home/styles.ts
@@ -1,0 +1,75 @@
+import { StyleSheet, Dimensions } from 'react-native';
+
+const { width, height } = Dimensions.get('window');
+
+const styles = StyleSheet.create({
+  outerContainer: {
+    flex: 1,
+    backgroundColor: '#fff',
+    paddingHorizontal: width * 0.1,
+  },
+  container: {
+    flex: 1,
+    justifyContent: 'flex-start',
+    backgroundColor: '#fff',
+    marginTop: height * 0.07,
+  },
+  title: {
+    fontSize: width * 0.05,
+    fontWeight: 'bold',
+    marginBottom: height * 0.04,
+    textAlign: 'center',
+  },
+  subtitle: {
+    fontSize: width * 0.05,
+    fontWeight: 'bold',
+    marginBottom: height * 0.03,
+    marginTop: height * 0.02,
+  },
+  card: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    padding: 15,
+    borderRadius: 10,
+    marginBottom: 15,
+    gap: 12,
+  },
+  icon: {
+    fontSize: 22,
+    textAlign: 'center',
+  },
+  cardText: {
+    flex: 1,
+  },
+  cardTitle: {
+    fontSize: 16,
+    fontWeight: 'bold',
+    color: '#000',
+  },
+  cardTime: {
+    fontSize: 14,
+    color: '#000',
+  },
+  cardIcon: {
+    width: 40,
+    height: 40,
+    borderRadius: 25,
+    backgroundColor: '#FFF',
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginRight: 16,
+  },
+  addButton: {
+    position: 'absolute',
+    bottom: 30,
+    right: 30,
+  },
+  cardImage: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    resizeMode: 'cover',
+  },
+});
+
+export default styles;
