+++ src/services/firebaseConfig.tsx
@@ -1,22 +1,15 @@
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
 import { getFirestore, collection,getDocs, query, where } from 'firebase/firestore';
-import {
-  FIREBASE_API_KEY,
-  FIREBASE_AUTH_DOMAIN,
-  FIREBASE_PROJECT_ID,
-  FIREBASE_STORAGE_BUCKET,
-  FIREBASE_MESSAGING_SENDER_ID,
-  FIREBASE_APP_ID
-} from '@env';
+import { getStorage } from "firebase/storage";
 
 const firebaseConfig = {
-  apiKey: FIREBASE_API_KEY,
-  authDomain: FIREBASE_AUTH_DOMAIN,
-  projectId: FIREBASE_PROJECT_ID,
-  storageBucket: FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: FIREBASE_MESSAGING_SENDER_ID,
-  appId: FIREBASE_APP_ID,
+  apiKey: "AIzaSyC3i1Sv8ELJLYQD8hEKxaNPMaPR1NYJYY8",
+  authDomain: "lembremed-f7ec3.firebaseapp.com",
+  projectId: "lembremed-f7ec3",
+  storageBucket: "lembremed-f7ec3.firebasestorage.app",
+  messagingSenderId: "46126987899",
+  appId: "1:46126987899:web:5484365fe25057749300de"
 };
 
 const app = initializeApp(firebaseConfig);
